moved some items into utils package
