+++ contracts/governance/src/proposal_state.rs
@@ -1,22 +1,22 @@
 use cosmwasm_std::{Storage, Uint128};
 use cosmwasm_storage::{
     bucket,
-    bucket_read,
-    singleton,
-    singleton_read,
     Bucket,
+    bucket_read,
     ReadonlyBucket,
     ReadonlySingleton,
+    singleton,
     Singleton,
+    singleton_read,
 };
 use secret_toolkit::snip20::batch::SendAction;
 use shade_protocol::{
-    generic_response::ResponseStatus,
     governance::{
         proposal::{Proposal, ProposalStatus},
         vote::VoteTally,
     },
 };
+use shade_protocol::utils::generic_response::ResponseStatus;
 
 // Proposals
 pub static PROPOSAL_KEY: &[u8] = b"proposals";
