+++ packages/network_integration/src/contract_helpers/stake.rs
@@ -1,6 +1,6 @@
 use crate::{
     contract_helpers::{governance::init_contract, minter::get_balance},
-    utils::{print_contract, print_header, ACCOUNT_KEY, GAS, STAKING_FILE},
+    utils::{ACCOUNT_KEY, GAS, print_contract, print_header, STAKING_FILE},
 };
 use cosmwasm_std::{HumanAddr, Uint128};
 use secretcli::{
@@ -8,8 +8,9 @@ use secretcli::{
     secretcli::{query_contract, test_contract_handle},
 };
 use serde_json::Result;
-use shade_protocol::{asset::Contract, snip20, staking};
+use shade_protocol::{snip20, staking};
 use std::{thread, time, time::UNIX_EPOCH};
+use shade_protocol::utils::asset::Contract;
 
 pub fn setup_staker(
     governance: &NetContract,
