+++ packages/shade_protocol/src/airdrop/mod.rs
@@ -5,13 +5,13 @@ use crate::{
         account::{AccountPermit, AddressProofPermit},
         claim_info::RequiredTask,
     },
-    asset::Contract,
-    generic_response::ResponseStatus,
 };
 use cosmwasm_std::{Binary, HumanAddr, Uint128};
 use schemars::JsonSchema;
 use secret_toolkit::utils::{HandleCallback, InitCallback, Query};
 use serde::{Deserialize, Serialize};
+use crate::utils::asset::Contract;
+use crate::utils::generic_response::ResponseStatus;
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct Config {
