+++ packages/shade_protocol/src/governance/mod.rs
@@ -1,11 +1,12 @@
 pub mod proposal;
 pub mod vote;
 
-use crate::{asset::Contract, generic_response::ResponseStatus};
+use crate::utils::generic_response::ResponseStatus;
 use cosmwasm_std::{Binary, HumanAddr, Uint128};
 use schemars::JsonSchema;
 use secret_toolkit::utils::{HandleCallback, InitCallback, Query};
 use serde::{Deserialize, Serialize};
+use crate::utils::asset::Contract;
 
 // This is used when calling itself
 pub const GOVERNANCE_SELF: &str = "SELF";
