+++ packages/shade_protocol/src/lib.rs
@@ -1,10 +1,7 @@
-// Helper libraries
-pub mod asset;
 pub mod band;
-pub mod generic_response;
-pub mod math;
 pub mod secretswap;
 pub mod snip20;
+pub mod utils;
 
 // Protocol init libraries
 pub mod airdrop;
