+++ packages/shade_protocol/src/mint.rs
@@ -1,10 +1,11 @@
-use crate::{asset::Contract, generic_response::ResponseStatus};
+use crate::utils::generic_response::ResponseStatus;
 use cosmwasm_std::{Binary, HumanAddr, Uint128};
 use schemars::JsonSchema;
 use secret_toolkit::utils::{HandleCallback, InitCallback, Query};
 #[cfg(test)]
 use secretcli::secretcli::{TestHandle, TestInit, TestQuery};
 use serde::{Deserialize, Serialize};
+use crate::utils::asset::Contract;
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct MintConfig {
