+++ packages/shade_protocol/src/oracle.rs
@@ -1,4 +1,4 @@
-use crate::{asset::Contract, generic_response::ResponseStatus, snip20::Snip20Asset};
+use crate::snip20::Snip20Asset;
 use cosmwasm_std::{HumanAddr, Uint128};
 use schemars::JsonSchema;
 use secret_toolkit::utils::{HandleCallback, InitCallback, Query};
@@ -6,6 +6,8 @@ use serde::{Deserialize, Serialize};
 
 #[cfg(test)]
 use secretcli::secretcli::{TestHandle, TestInit, TestQuery};
+use crate::utils::asset::Contract;
+use crate::utils::generic_response::ResponseStatus;
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct SswapPair {
