+++ packages/shade_protocol/src/scrt_staking.rs
@@ -1,8 +1,9 @@
-use crate::{asset::Contract, generic_response::ResponseStatus};
+use crate::utils::generic_response::ResponseStatus;
 use cosmwasm_std::{Binary, Decimal, FullDelegation, HumanAddr, Uint128, Validator};
 use schemars::JsonSchema;
 use secret_toolkit::utils::{HandleCallback, InitCallback, Query};
 use serde::{Deserialize, Serialize};
+use crate::utils::asset::Contract;
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
