+++ packages/shade_protocol/src/snip20.rs
@@ -1,4 +1,4 @@
-use crate::asset::Contract;
+use crate::utils::asset::Contract;
 use cosmwasm_std::{Binary, HumanAddr, Querier, StdResult, Uint128};
 use schemars::JsonSchema;
 use secret_toolkit::{
