+++ packages/shade_protocol/src/treasury.rs
@@ -1,4 +1,4 @@
-use crate::{asset::Contract, generic_response::ResponseStatus};
+use crate::utils::generic_response::ResponseStatus;
 use cosmwasm_std::{Binary, Decimal, HumanAddr, Uint128};
 use schemars::JsonSchema;
 use secret_toolkit::{
@@ -6,6 +6,7 @@ use secret_toolkit::{
     utils::{HandleCallback, InitCallback, Query},
 };
 use serde::{Deserialize, Serialize};
+use crate::utils::asset::Contract;
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct Config {
