+++ packages/shade_protocol/src/utils/asset.rs
