+++ packages/shade_protocol/src/utils/generic_response.rs
